# Cache asset graph, bundle graph, and packager/optimizer results