--- a/packages/core/core/src/BundlerRunner.js
+++ b/packages/core/core/src/BundlerRunner.js
@@ -14,7 +14,9 @@ import MainAssetGraph from './public/MainAssetGraph';
 import {Bundle, NamedBundle} from './public/Bundle';
 import AssetGraphBuilder from './AssetGraphBuilder';
 import {report} from './ReporterRunner';
-import {normalizeSeparators, unique} from '@parcel/utils';
+import {normalizeSeparators, unique, md5FromObject} from '@parcel/utils';
+import Cache from '@parcel/cache';
+import {localResolve} from '@parcel/local-require';
 
 type Opts = {|
   options: ParcelOptions,
@@ -24,10 +26,12 @@ type Opts = {|
 export default class BundlerRunner {
   options: ParcelOptions;
   config: ParcelConfig;
+  cache: Cache;
 
   constructor(opts: Opts) {
     this.options = opts.options;
     this.config = opts.config;
+    this.cache = new Cache(this.options.outputFS, this.options.cacheDir);
   }
 
   async bundle(graph: AssetGraph): Promise<InternalBundleGraph> {
@@ -35,6 +39,15 @@ export default class BundlerRunner {
       type: 'buildProgress',
       phase: 'bundling'
     });
+
+    let cacheKey;
+    if (this.options.cache !== false) {
+      cacheKey = await this.getCacheKey(graph);
+      let cachedBundleGraph = await this.cache.get(cacheKey);
+      if (cachedBundleGraph) {
+        return cachedBundleGraph;
+      }
+    }
 
     let bundler = await this.config.getBundler();
 
@@ -47,7 +60,26 @@ export default class BundlerRunner {
     await this.nameBundles(bundleGraph);
     await this.applyRuntimes(bundleGraph);
 
+    if (cacheKey != null) {
+      await this.cache.set(cacheKey, bundleGraph);
+    }
+
     return bundleGraph;
+  }
+
+  async getCacheKey(assetGraph: AssetGraph) {
+    let bundler = this.config.bundler;
+    let [, resolvedPkg] = await localResolve(
+      `${bundler}/package.json`,
+      `${this.config.filePath}/index` // TODO: is this right?
+    );
+
+    let version = nullthrows(resolvedPkg).version;
+    return md5FromObject({
+      bundler,
+      version,
+      hash: assetGraph.getHash()
+    });
   }
 
   async nameBundles(bundleGraph: InternalBundleGraph): Promise<void> {
@@ -142,7 +174,8 @@ export default class BundlerRunner {
     let bundle = node.value;
 
     for (let {code, filePath, dependency} of runtimeAssets) {
-      let builder = new AssetGraphBuilder({
+      let builder = new AssetGraphBuilder();
+      await builder.init({
         options: this.options,
         config: this.config,
         assetRequest: {