--- a/packages/core/core/src/PackagerRunner.js
+++ b/packages/core/core/src/PackagerRunner.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type {ParcelOptions, Blob, FilePath} from '@parcel/types';
+import type {ParcelOptions, Blob, FilePath, BundleResult} from '@parcel/types';
 import type SourceMap from '@parcel/source-map';
 import type {Bundle as InternalBundle} from './types';
 import type ParcelConfig from './ParcelConfig';
@@ -9,13 +9,20 @@ import type {FileSystem, FileOptions} from '@parcel/fs';
 
 import {Readable} from 'stream';
 import invariant from 'assert';
-import {urlJoin} from '@parcel/utils';
+import {
+  urlJoin,
+  md5FromObject,
+  md5FromString,
+  blobToStream
+} from '@parcel/utils';
 import {NamedBundle} from './public/Bundle';
 import nullthrows from 'nullthrows';
 import path from 'path';
 import url from 'url';
 import {report} from './ReporterRunner';
 import {BundleGraph} from './public/BundleGraph';
+import Cache from '@parcel/cache';
+import {localResolve} from '@parcel/local-require';
 
 type Opts = {|
   config: ParcelConfig,
@@ -27,23 +34,35 @@ export default class PackagerRunner {
   options: ParcelOptions;
   distDir: FilePath;
   distExists: Set<FilePath>;
+  cache: Cache;
 
   constructor({config, options}: Opts) {
     this.config = config;
     this.options = options;
     this.distExists = new Set();
+    this.cache = new Cache(options.outputFS, options.cacheDir);
   }
 
   async writeBundle(bundle: InternalBundle, bundleGraph: InternalBundleGraph) {
     let {inputFS, outputFS} = this.options;
     let start = Date.now();
-    let packaged = await this.package(bundle, bundleGraph);
-    let {contents, map} = await this.optimize(
-      bundle,
-      packaged.contents,
-      packaged.map
-    );
 
+    let result, cacheKey;
+    if (this.options.cache !== false) {
+      cacheKey = await this.getCacheKey(bundle);
+      result = await this.readFromCache(cacheKey);
+    }
+
+    if (!result) {
+      let packaged = await this.package(bundle, bundleGraph);
+      result = await this.optimize(bundle, packaged.contents, packaged.map);
+
+      if (cacheKey != null) {
+        await this.writeToCache(cacheKey, result);
+      }
+    }
+
+    let {contents, map} = result;
     let filePath = nullthrows(bundle.filePath);
     let dir = path.dirname(filePath);
     if (!this.distExists.has(dir)) {
@@ -124,7 +143,7 @@ export default class PackagerRunner {
   async package(
     internalBundle: InternalBundle,
     bundleGraph: InternalBundleGraph
-  ): Promise<{|contents: Blob, map?: ?SourceMap|}> {
+  ): Promise<BundleResult> {
     let bundle = new NamedBundle(internalBundle);
     report({
       type: 'buildProgress',
@@ -156,7 +175,7 @@ export default class PackagerRunner {
     internalBundle: InternalBundle,
     contents: Blob,
     map?: ?SourceMap
-  ): Promise<{|contents: Blob, map?: ?SourceMap|}> {
+  ): Promise<BundleResult> {
     let bundle = new NamedBundle(internalBundle);
     let optimizers = await this.config.getOptimizers(bundle.filePath);
     if (!optimizers.length) {
@@ -180,6 +199,56 @@ export default class PackagerRunner {
     }
 
     return optimized;
+  }
+
+  async getCacheKey(bundle: InternalBundle) {
+    let filePath = nullthrows(bundle.filePath);
+    let packager = this.config.getPackagerName(filePath);
+    let optimizers = this.config.getOptimizerNames(filePath);
+    let deps = Promise.all(
+      [packager, ...optimizers].map(async pkg => {
+        let [, resolvedPkg] = await localResolve(
+          `${pkg}/package.json`,
+          `${this.config.filePath}/index` // TODO: is this right?
+        );
+
+        let version = nullthrows(resolvedPkg).version;
+        return [pkg, version];
+      })
+    );
+
+    // TODO: add third party configs to the cache key
+    let {minify, scopeHoist, sourceMaps} = this.options;
+    return md5FromObject({
+      deps,
+      opts: {minify, scopeHoist, sourceMaps},
+      hash: bundle.assetGraph.getHash()
+    });
+  }
+
+  async readFromCache(cacheKey: string): Promise<?BundleResult> {
+    let contentKey = md5FromString(`${cacheKey}:content`);
+    let mapKey = md5FromString(`${cacheKey}:map`);
+
+    let contentExists = await this.cache.blobExists(contentKey);
+    if (!contentExists) {
+      return null;
+    }
+
+    return {
+      contents: this.cache.getStream(contentKey),
+      map: await this.cache.get(mapKey)
+    };
+  }
+
+  async writeToCache(cacheKey: string, result: BundleResult) {
+    let contentKey = md5FromString(`${cacheKey}:content`);
+
+    await this.cache.setStream(contentKey, blobToStream(result.contents));
+    if (result.map) {
+      let mapKey = md5FromString(`${cacheKey}:map`);
+      await this.cache.set(mapKey, result.map);
+    }
   }
 }