--- a/packages/core/core/src/Parcel.js
+++ b/packages/core/core/src/Parcel.js
@@ -88,7 +88,8 @@ export default class Parcel {
       options: resolvedOptions
     });
 
-    this.#assetGraphBuilder = new AssetGraphBuilder({
+    this.#assetGraphBuilder = new AssetGraphBuilder();
+    await this.#assetGraphBuilder.init({
       options: resolvedOptions,
       config,
       entries: resolvedOptions.entries,
@@ -106,13 +107,18 @@ export default class Parcel {
   }
 
   async run(): Promise<IBundleGraph> {
+    let startTime = Date.now();
     if (!this.#initialized) {
       await this.init();
     }
 
-    let result = await this.build();
+    let result = await this.build(startTime);
 
     let resolvedOptions = nullthrows(this.#resolvedOptions);
+    if (result.type === 'buildSuccess') {
+      await this.#assetGraphBuilder.writeToCache();
+    }
+
     if (resolvedOptions.killWorkers !== false) {
       await this.#farm.end();
     }
@@ -172,13 +178,12 @@ export default class Parcel {
     };
   }
 
-  async build(): Promise<BuildEvent> {
+  async build(startTime: number = Date.now()): Promise<BuildEvent> {
     try {
       this.#reporterRunner.report({
         type: 'buildStart'
       });
 
-      let startTime = Date.now();
       let {assetGraph, changedAssets} = await this.#assetGraphBuilder.build();
       dumpGraphToGraphViz(assetGraph, 'MainAssetGraph');
 
@@ -265,11 +270,7 @@ export default class Parcel {
     invariant(this.#watcherSubscription == null);
 
     let resolvedOptions = nullthrows(this.#resolvedOptions);
-    let targetDirs = resolvedOptions.targets.map(target => target.distDir);
-    let vcsDirs = ['.git', '.hg'].map(dir =>
-      path.join(resolvedOptions.projectRoot, dir)
-    );
-    let ignore = [resolvedOptions.cacheDir, ...targetDirs, ...vcsDirs];
+    let opts = this.#assetGraphBuilder.getWatcherOptions();
 
     return watcher.subscribe(
       resolvedOptions.projectRoot,
@@ -293,7 +294,7 @@ export default class Parcel {
           }
         }
       },
-      {ignore}
+      opts
     );
   }
 }