… packages/core/fs/src/NodeFS.js
… packages/core/source-map/src/SourceMap.js
… packages/core/types/index.js
… packages/core/utils/src/stream.js
… packages/namers/default/src/DefaultNamer.js